requirements/base.txt
Django==2.0.7
djangorestframework==3.8.2
django-model-utils-3.1.2
psycopg2-binary==2.7.5
python-decouple==3.1
dj-database-url==0.5.0
Unipath==1.1
